feat(analytics): extract generalized loader
