--- projects/client/src/lib/features/analytics/AnalyticsProps.ts
@@ -1,5 +1,0 @@
-import type { AnalyticsEngine } from './_internal/AnalyticsEngine.ts';
-
-export type AnalyticsProps = {
-  onload: (engine: AnalyticsEngine) => void;
-};
