+++ projects/client/src/lib/features/analytics/_internal/loader/AnalyticsProps.ts
@@ -1,0 +1,3 @@
+export type AnalyticsProps = {
+  onload: (userId: string | Nil) => void;
+};
