+++ projects/client/src/lib/features/analytics/_internal/loader/AnonymousAnalytics.svelte
@@ -2,9 +2,7 @@
   import { isSupported } from "firebase/analytics";
   import { onMount } from "svelte";
   import type { AnalyticsProps } from "./AnalyticsProps";
-  import { useAnalytics } from "./useAnalytics";
 
-  const analytics = useAnalytics();
   const { onload }: AnalyticsProps = $props();
 
   onMount(async () => {
@@ -12,7 +10,6 @@
       return;
     }
 
-    analytics.setUserId(null);
-    onload(analytics);
+    onload(null);
   });
 </script>
